requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-MySQLdb
Flask-SQLAlchemy
Werkzeug
python-dotenv
spacy
scikit-learn
pandas
nltk
email_validator